separate add VM feature by CQRS
