+++ conoha_client/features/vm/repo/test_command.py
@@ -1,9 +1,6 @@
 """command tests."""
 from __future__ import annotations
 
-import json
-from ipaddress import IPv4Address
-from pathlib import Path
 from typing import TYPE_CHECKING
 from uuid import UUID, uuid4
 
@@ -13,9 +10,8 @@ from conoha_client.features._shared.conftest import prepare
 from conoha_client.features._shared.endpoints.endpoints import Endpoints
 from conoha_client.features.vm.errors import (
     NotFlavorProvidesError,
-    NotFoundAddedVMError,
 )
-from conoha_client.features.vm.repo.command import AddVMCommand, find_added
+from conoha_client.features.vm.repo.command import AddVMCommand
 
 if TYPE_CHECKING:
     from requests_mock import Mocker
@@ -49,11 +45,10 @@ def test_add_vm() -> None:
         flavor_id=uuid4(),
         image_id=uuid4(),
         admin_pass="xxx",  # noqa: S106
-        sshkey_name=None,
-        post=mock_post,
+        dep=mock_post,
     )
 
-    added = cmd()
+    added = cmd(None)
     assert added.vm_id == UUID(VM_ID)
 
 
@@ -61,7 +56,7 @@ def test_invalid_add_vm(
     requests_mock: Mocker,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    """Invalid case."""
+    """Invalid case error code 400. 課金に関わるから慎重."""
     prepare(requests_mock, monkeypatch)
     requests_mock.post(Endpoints.COMPUTE.tenant_id_url("servers"), status_code=400)
 
@@ -69,32 +64,6 @@ def test_invalid_add_vm(
         flavor_id=uuid4(),
         image_id=uuid4(),
         admin_pass="xxx",  # noqa: S106
-        sshkey_name=None,
     )
     with pytest.raises(NotFlavorProvidesError):
-        cmd()
-
-
-def test_find_ipv4() -> None:
-    """Test."""
-    expected = IPv4Address("157.7.78.59")  # fixutureのipv4
-
-    def mock_get_servers() -> list[object]:
-        """list_imageのモック."""
-        p = Path(__file__).resolve().parent / "fixture_servers20231006.json"
-        return json.loads(p.read_text())
-
-    added = find_added(UUID(VM_ID), dep=mock_get_servers)
-    assert added.ipv4 == expected
-
-
-def test_invalid_find_ipv4() -> None:
-    """Invalid case."""
-
-    def mock_get_servers() -> list[object]:
-        """2個目しか契約中VMが見つからない."""
-        p = Path(__file__).resolve().parent / "fixture_servers20231006.json"
-        return [json.loads(p.read_text())[1]]
-
-    with pytest.raises(NotFoundAddedVMError):
-        find_added(UUID(VM_ID), dep=mock_get_servers)
+        cmd(None)
